+++ CHANGELOG.md
@@ -15,7 +15,8 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 ## Changed
 
 * Configs are deployed using docker's `configs` section. 
-* NOC's config are mounted in /etc/noc/settings.yml
+* NOC's config are mounted in /etc/noc/settings.yml.
+* Move postgres' password to secrets.
 
 ## 0.11.0 - 2025-09-12